Remove static energy bilance
This information is redundant due to the load profile. Hence the high memory demand of the
model, storing redundant information should be avoided. This change moves the energy
calculation out of the load profile from the init phase to the simulation runtime. So the outcome
of this change is a lower memory demand at cost of a slightly higher runtime.