Init python version of EnSySim
A new python based implementation of the model allows
in combination with numba a performant and well structured
model approach. Due to the vectorisation needed to keep
calculation times in matlab tolerable,
there where compromises necessary in terms of the model structure.
This new approach will allow a well suited structure as well as
faster simulation runs.